code to output file with dtwdistance between 2 main channels
The function to get the DTW distance is implemented. A function to return the list of points in order together with the cost for each point for an optimal solution using DTW are both implemented.
There is a function to get the best matching channel but this one is not complete.

There is still some code which was used for testing and needs to be cleaned